[refactoring] 파일명 변경 및 ICommand.h 분리
